requirements.txt
pandas==1.2.3
cassandra-driver==3.25.0
black==20.8b1
